Collect JSON schemas in betty.json_schemas
