+++ betty/entities/note.py
@@ -14,9 +14,9 @@ from betty.entity.association import (
     ToZeroOrOneAssociate,
 )
 from betty.entity.has_links import HasLinks
+from betty.json_schemas.static_translations import StaticTranslationsSchema
 from betty.locale.localizable.gettext import _, ngettext
 from betty.locale.localizable.linked_data import dump_linked_data
-from betty.locale.localizable.static.schema import StaticTranslationsSchema
 from betty.privacy import Privacy
 
 if TYPE_CHECKING:
