+++ betty/entity/association.py
@@ -16,8 +16,8 @@ from betty.entity.collection import (
 )
 from betty.entity.collection.multiple import MultipleTypesEntityCollection
 from betty.entity.collection.single import SingleTypeEntityCollection
-from betty.entity.schema import ToManySchema, ToZeroOrOneSchema
 from betty.importlib import fully_qualified_name, import_any
+from betty.json_schemas.entity_association import ToManySchema, ToZeroOrOneSchema
 from betty.linked_data import LinkedDataDumper
 from betty.locale.localizable import resolve_localizable
