+++ betty/jobs/generate_json_error_responses.py
@@ -10,9 +10,9 @@ from typing import TYPE_CHECKING, final, override
 
 from betty.file import write
 from betty.job import Job
+from betty.json_schemas.project import ProjectSchema
 from betty.locale.localizable.gettext import _
 from betty.locale.localize import default_localizer
-from betty.project.schema import ProjectSchema
 
 if TYPE_CHECKING:
     from betty.job.scheduler import Scheduler
