+++ betty/jobs/generate_json_schema.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, final, override
 
 from betty.file import write
 from betty.job import Job
-from betty.project.schema import ProjectSchema
+from betty.json_schemas.project import ProjectSchema
 
 if TYPE_CHECKING:
     from betty.job.scheduler import Scheduler
