+++ betty/json_schemas/__init__.py
@@ -1,0 +1,3 @@
+"""
+Concrete, reusable JSON Schemas.
+"""
