+++ betty/json_schemas/static_translations.py
