+++ betty/tests/jobs/test_generate_openapi.py
@@ -1,7 +1,7 @@
 import json
 
 from betty.jobs.generate_openapi import GenerateOpenapi
-from betty.openapi.schema import SpecificationSchema
+from betty.json_schemas.openapi import OpenapiSchema
 from betty.project import Project
 from betty.test_utils.job import do
 
@@ -13,4 +13,4 @@ class TestGenerateOpenapi:
         with open(
             isolated_project.www_directory / "api" / "index.json", encoding="utf-8"
         ) as f:
-            SpecificationSchema().validate(json.loads(f.read()))
+            OpenapiSchema().validate(json.loads(f.read()))
