+++ betty/tests/json_schemas/test_json_schema.py
@@ -2,12 +2,12 @@ from typing import override
 
 import pytest
 
-from betty.openapi.schema import SpecificationSchema
+from betty.json_schemas.json_schema import JsonSchemaSchema
 from betty.test_utils.json_schema import SchemaTestBase, SchemaTestBaseSut
 
 
-class TestSpecificationSchema(SchemaTestBase):
+class TestJsonSchemaSchema(SchemaTestBase):
     @override
     @pytest.fixture
     def sut_data(self) -> SchemaTestBaseSut:
-        return (SpecificationSchema(), [], [])
+        return (JsonSchemaSchema(), [], [])
