+++ betty/tests/json_schemas/test_media_type.py
@@ -4,7 +4,7 @@ from typing import override
 
 import pytest
 
-from betty.media_type.schema import MediaTypeSchema
+from betty.json_schemas.media_type import MediaTypeSchema
 from betty.test_utils.json_schema import SchemaTestBase, SchemaTestBaseSut
