+++ betty/tests/json_schemas/test_openapi.py
@@ -1,0 +1,13 @@
+from typing import override
+
+import pytest
+
+from betty.json_schemas.openapi import OpenapiSchema
+from betty.test_utils.json_schema import SchemaTestBase, SchemaTestBaseSut
+
+
+class TestOpenapiSchema(SchemaTestBase):
+    @override
+    @pytest.fixture
+    def sut_data(self) -> SchemaTestBaseSut:
+        return (OpenapiSchema(), [], [])
