+++ betty/tests/locale/localizable/test_linked_data.py
@@ -1,9 +1,9 @@
 import pytest
 
+from betty.json_schemas.static_translations import StaticTranslationsSchema
 from betty.locale.localizable import ShorthandStaticTranslations
 from betty.locale.localizable.linked_data import dump_linked_data
 from betty.locale.localizable.static import StaticTranslations
-from betty.locale.localizable.static.schema import StaticTranslationsSchema
 from betty.locale.localize import default_localizer
 from betty.portable import PortableMapping
 from betty.test_utils.conftest import AssertLinkedDataDump
