+++ betty/tests/locale/localizable/test_static.py
