+++ betty/tests/test_openapi.py
@@ -1,7 +1,7 @@
 import pytest
 
+from betty.json_schemas.openapi import OpenapiSchema
 from betty.openapi import Specification
-from betty.openapi.schema import SpecificationSchema
 from betty.test_utils.conftest import IsolatedProjectFactory
 
 
@@ -19,4 +19,4 @@ class TestSpecification:
         async with isolated_project_factory(clean_urls=clean_urls) as project:
             sut = Specification(project)
             specification = await sut.build()
-        SpecificationSchema().validate(specification)
+        OpenapiSchema().validate(specification)
